renderer: do explicit uint casting in lighttile_fp.glsl

+++ src/engine/renderer/glsl_source/lighttile_fp.glsl
@@ -59,7 +59,7 @@ uniform sampler2D u_DepthMap;
 uniform int u_lightLayer;
 uniform vec3 u_zFar;
 
-const int lightsPerLayer = 16;
+const uint lightsPerLayer = 16u;
 
 #define idxs_t uvec4
 
@@ -67,8 +67,8 @@ DECLARE_OUTPUT(uvec4)
 
 // 8 bits per light ID
 void pushIdxs( in uint idx, in uint count, inout uvec4 idxs ) {
-	idxs[count / 4] <<= 8;
-	idxs[count / 4] |= idx & 0xFFu;
+	idxs[count / 4u] <<= 8u;
+	idxs[count / 4u] |= idx & 0xFFu;
 }
 
 #define exportIdxs( x ) outputColor = ( x )
@@ -114,13 +114,13 @@ void main() {
 
 	idxs_t idxs = uvec4( 0, 0, 0, 0 );
 
-	uint lightCount = 0;
+	uint lightCount = 0u;
 
 	/* Dynamic lights are put into 4 layers of a 3D texture. Since checking if we already added some light is infeasible,
 	only process 1 / 4 of different lights for each layer, extra lights going into the last layer. This can fail to add some lights
 	if 1 / 4 of all lights is more than the amount of lights that each layer can hold (16). To fix this, we'd need to either do this on CPU
 	or use compute shaders with atomics so we can have a variable amount of lights for each tile. */
-	for( uint i = u_lightLayer; i < u_numLights; i += NUM_LIGHT_LAYERS ) {
+	for( uint i = uint( u_lightLayer ); i < uint( u_numLights ); i += uint( NUM_LIGHT_LAYERS ) ) {
 		Light l = GetLight( i );
 		vec3 center = ( u_ModelMatrix * vec4( l.center, 1.0 ) ).xyz;
 		float radius = max( 2.0 * l.radius, 2.0 * 32.0 ); // Avoid artifacts with weak light sources
@@ -136,7 +136,7 @@ void main() {
 		if( radius > 0.0 ) {
 			/* Light IDs are stored relative to the layer
 			Add 1 because 0 means there's no light */
-			pushIdxs( ( i / NUM_LIGHT_LAYERS ) + 1, lightCount, idxs );
+			pushIdxs( ( i / uint( NUM_LIGHT_LAYERS ) ) + 1u, lightCount, idxs );
 			lightCount++;
 
 			if( lightCount == lightsPerLayer ) {
